LayoutTests/http/tests/site-isolation/page-visibility-onvisibilitychange-in-iframe.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->

<!DOCTYPE html>
<html>
<body onload="startTest()">
<iframe id="iframeTest" src="http://localhost:8000/site-isolation/resources/page-visibility-iframe-send-onvisibility-change-back.html"></iframe>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("This test checks the onvisibilitychange event handler attribute in an iframe.");

var jsTestIsAsync = true;

function startTest() {
    if (!window.testRunner)
        return;
    
    window.addEventListener("message", function(event) {
        shouldBeEqualToString("event.data.text", "visibility change triggered");
        finishJSTest();
    }, false);

    if (window.testRunner)
        testRunner.setPageVisibility("hidden");

    testRunner.waitUntilDone();
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
